frontend: Clean up installation steps code
Rename WaitingLi to Step and move to tf-poweron.jsx.
Clean up steps CSS a bit.
