+++ installer/frontend/components/ui.jsx
@@ -606,29 +606,6 @@ export const PrivateKeyArea = (props) => {
   return <FileArea {...areaProps} />;
 };
 
-export const WaitingLi = ({pending, done, error, cancel, children, substep}) => {
-  const progressClasses = classNames({
-    'wiz-launch-progress__step': !substep,
-    'wiz-launch-progress__substep': substep,
-    'wiz-pending-fg': pending,
-    'wiz-error-fg': error,
-    'wiz-success-fg': done && !error,
-    'wiz-cancel-fg': !done && !error && cancel,
-    'wiz-running-fg': !done && !error && !cancel && !pending,
-  });
-  const iconClasses = classNames('fa', 'fa-fw', {
-    'fa-circle-o': pending,
-    'fa-exclamation-circle': error,
-    'fa-check-circle': done && !error,
-    'fa-ban': !done && !error && cancel,
-    'fa-spin fa-circle-o-notch': !done && !error && !cancel && !pending,
-  });
-
-  return <li className={progressClasses}>
-    {!substep && <i className={iconClasses} style={{margin: '0 3px 0 -3px'}}></i>}{children}
-  </li>;
-};
-
 export class AsyncSelect extends React.Component {
   componentDidMount () {
     const { onChange, onRefresh, value } = this.props;
